refactor(shared): move schedule utils and PeriodTimes to :shared
